requirements.txt
slack-bolt==1.19.1
python-dotenv==1.0.1
langchain==0.2.14


